Time: 3 ms (54.24%) | Memory: 6.4 MB (25.42%) - LeetSync

+++ problems/492-construct-the-rectangle/construct-the-rectangle.c
@@ -3,7 +3,7 @@
  */
 #include <stdlib.h>
 int* constructRectangle(int area, int* returnSize) {
-    int j=(int)sqrt(area);
+    int j=sqrt(area);
     while(area%j!=0)
         j-=1;
     int* a=(int*)malloc(2*sizeof(int));
